Misc Updates
Modified the spacing of the years of service on each staff member, modified the homepage slider slightly to be more mobile friendly. Changed the nav-buttons to have a slight border to make them stand out in front of the images in the slider. Modified the nav buttons to contract a little when the screen resolution is smaller than 1920x1080 to avoid strange overflows. Modified the officer info boxes to be divs that are displayed inline-block so that they will wrap nicely when the screen resolution is smaller. Ditto for the social media containers on the homepage.