Several changes
Added a new outreach page. Changed manage user permissions to be under manage rather than admin, Fixed some Error 500 errors when trying to update things from any of the manage pages, changed the pictures on the homepage to be the size of the viewport rather than 1920x1080 and set their max-width and max-height to 1920 and 1080 respectively.